backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
opencv-python==4.9.0.80
numpy==1.26.4
scipy==1.13.0
Pillow==10.3.0
pillow-heif==0.16.0
PyWavelets==1.6.0
paddlepaddle==2.6.1
paddleocr==2.7.3
spacy==3.7.4
transformers==4.41.1
torch==2.3.0
scikit-learn==1.5.0
nltk==3.8.1
python-dotenv==1.0.1
werkzeug==3.0.3
matplotlib==3.9.0
